Introduce assert:follows-standard-deprecation tag
The assert:follows-standard-deprecation tag defines a standard
deprecation policy for user-visible and ops-visible features and
APIs in OpenStack services. Project teams may assert that they
follow at least this policy (and then commit to it).

This is an essential bit of information to communicate to our users,
and a key factor in assessing a given project maturity.

Co-Authored-By: Zane Bitter <zbitter@redhat.com>

Change-Id: Ic241e93e83422deaf971238a2049160b75bd0bf3