Add new repository for stand-alone TOSCA Parser
The heat-translator project currently contains two components - Translator and
TOSCA Parser. In order to use them independently we want to split them to
their own repo so that they can be offered as a stand-alone PyPI packages. I
have discussed it with the Heat PTL and he is okay with splitting the repos.

Change-Id: Ie02baa22737474f4402f17906ec1e7ff100357d2