Temporarily add I18n ATCs
Because there is no automatic method to recognize the active contributors
of I18n team, we have to use this temporary way to add I18n ATCs,
to get the vote right and discount code for translators.
After we officially move to Zanata, there could be an automatic way
to get the I18n ATC list. At that time, the naming list will be removed
from here.

Change-Id: I7be23b2bca64a9d78a5be18941e0398171422936