Remove MagnetoDB from OpenStack
Over the liberty cycle MagnetoDB had 6 commits, 3 of which were
procedural. It had 14 reviews. It now has no PTL candidate. While
the effort on the project was appreciated, it is clear that the project
has joined burrow and melange in the realm of projects of yesteryear.

In addition to removing the project from the reference, the MagnetoDB repo
should receive a commit that deletes all the content and replaces it with a
README file that indicates the project has retired.

Change-Id: I6cc8de967bcbe021abb16384abb8faa32de1edc9