CloudKitty application for Big Tent
This is a request to add CloudKitty [1] to the Big Tent.

CloudKitty is a rating as a Service component aimed at easing
integration of OpenStack with existing billing systems. It consumes data
from metric systems (ceilometer for example) and with the help of rating
plugins and client defined rules, generates prices for the different
services.

The project is fully licensed under the Apache 2.0 license, and has been
actively promoted during the last 2 cycles including a presentation at
Vancouver summit, and another on stage presentation.

Stéphane Albert is currently acting as PTL. He has been working on the
project since the first POC shown at the OpenStack Summit Atlanta. We
plan on holding an official PTL election around the Tokyo summit.

All code changes are approved by the core members. CloudKitty code is
submitted and reviewed using the OpenStack infrastructure and gated
through gerrit[2][3][4].

We are starting to use both dev and operators mailing list in order to
give the community a way to provide feedback, keep up with our progress
and intentions.
Meanwhile we have been heavily using IRC for cooperation.
The #cloudkitty IRC channel on freenode is logged and the channel is
notified of such logging [5]. Community members, core developers and
users gather here to discuss about the future of the project as well as
questions regarding usage.
We have recently booked a slot on #openstack-meeting-3 in for our
regular meetings.

CloudKitty has been initiated by a single company but since then we have
received several patches, and reviews from exterior contributors both
originating from different companies and unaffiliated.
Actually the whole design of CloudKitty has been improved and decided
during regular IRC meetings that have been held by contibutors from many
companies in order to reflect their needs and usages.

We are aware that CloudKitty has been used in production by several
Cloud User (both Cloud Public or Private) for several months.

[1]https://wiki.openstack.org/wiki/CloudKitty
[2]https://review.openstack.org/#/q/project:stackforge/cloudkitty,n,z
[3]https://review.openstack.org/#/q/project:stackforge/python-cloudkittyclient,n,z
[4]https://review.openstack.org/#/q/project:stackforge/cloudkitty-dashboard,n,z
[5]http://eavesdrop.openstack.org/irclogs/%23cloudkitty/

Change-Id: I1864953db026832e8c40707e212db0ef179fe480