stackforge retirement: make workflow reqs+cla clearer
There has been some concern voiced over the ambiguity of the language
used in this resolution. Concerns were brought up about whether new or
existing contributors will now need to sign a CLA or alter workflow. It
is my understanding that this is not the case.

To make this clear, I propose adding this unambiguous writing to the
resolution.

Change-Id: Iac0f87678479e694350cccfb9bb87ce3bb946ff6