Note that the TC expects some project history.
The recent application by the Kosmos project [1] prompted a discussion
on whether or not a project should apply to the TC right away or if it
should have some amount of history of operating as an OpenStack project
first.  The consensus on that application was that we should wait and
let the project get going first.

This patch updates the new projects requirements to clarify this point.
As noted by ttx on [1], this guideline will henceforth be known as
"the Kosmos jurisprudence".

[1] https://review.openstack.org/#/c/223674/

Change-Id: Ifc964d09ea2107e3ecd95f9579543bcb77f457d0
Signed-off-by: Russell Bryant <rbryant@redhat.com>