remove naming timebox restriction
There is no reason to hold off on naming releases one we know their
location. And holding off on release naming is more likely to create
confusion where the dev community just picks a placeholder, uses it in
communication, and that placeholder is wrong later.

Especially as projects are now moving into a phase of planning for
multiple releases in advance for larger efforts, the current
restriction just adds confusion and cleanup work later.

Change-Id: Ife143447ff4a6d2d85260fd9d889853863eff272