Add a tool that allows visualizing the projects as a graph
This tool, when ran can create diagrams that show what
exactly the big-tent is and/or what the subprojects/deliverables
a project produces are.

For example, here is the way to create a output for ironic
and its deliverables:

$ python tools/universe_dot.py  reference/projects.yaml \
  ironic_universe.dot ironic;

$ neato -Tsvg ironic_universe.dot > ironic_universe.svg

Then view svg as is, or convert to png or other...

This requires pydot2 and pyyaml to be installed to work.

Change-Id: Idcd73f9e678e200d65eda4d1d54f1156fd18853b