Add senlin project to big tent
This is a request to add Senlin [1] to the Big Tent.

Senlin is designed to be a generic clustering service for OpenStack clouds
that makes the creation and management of homogeneous objects as groups an
easier task. Examples of objects include Heat stacks, Nova servers, etc.
Senlin provides an open framework that consists of:

 - an engine that manages clusters and nodes and the primitives to operate
   them as abstract objects;
 - profile plugins that enable senlin to talk to a specific OpenStack
   service for object CRUD operations.
 - policy plugins that capture rule sets which are checked/enforced when
   object operations are performed, e.g. placement, deletion, scaling, health
   and others.
 - trigger plugins that customize the external signals/alarms/events that can
   trigger operations on clusters/objects/profiles/policies;

The project was started in Nov, 2014 after the Paris summit as a standalone
service that will offload the auto-scaling service from Heat. The scope was
then expanded to include other policies for a comprehensive clustering or
collection service on OpenStack. The project was openly discussed with the Heat
team during Vancouver summit. Plan is being developed for migrating the
existing auto-scaling support in a way that reduce (or totally avoid) user land
changes.

Senlin is also working with other OpenStack projects to explore collaboration
opportunities. For example, team has worked with intern students to manage
container clusters under the Magnum context. The goal is to investigate how to
manage the VM groups and container clusters so that these two layers can be
scaled, load-balanced in accordance with each other. A demo is scheduled on
the Tokyo summit.

Senlin strictly follows the "4 opens" principle of the community.
The project is fully licensed under the Apache 2.0 license. All code changes
are submitted and reviewed using the OpenStack infrastructure and gated
through gerrit[2][3][4]. Team has been using the +2/+A process for patch
reviews. The current unit test coverage is 90% and functional gate is in place
helping ensure everything works together. The project strictly obeys the
coordinated project interfaces, including tox, pbr, global-requirements etc.
All dependencies of the project falls within the range of global requirements.
The gating now runs pep8, py27, py34, docs tasks and optionally a functional
test.

The PTL was chosen by election to be Qiming Teng during Mitaka cycle. The PTL
serves as liason between the Senlin community and other projects to coordinate
the interactions with for example python-openstacksdk, heat, ceilometer, etc.
Team plans to excercise a PTL rotation among core contributors.

Project team has been using openstack-dev mailing list and freenode IRC
channel #senlin heavily for cooperation. The #senlin channel is logged [5] and
the logging is notified in the channel. Project team has weekly IRC meeting on
channel #openstack-meeting on 1300 UTC every Tuesday [6] to accommodate
contributors from east Asia and America.

Senlin has a well-documented, reasonably stable API [7] ready to be merged by
upstream project. It has documents for beginner users [8] and new developers
[9]. Developer team has been answering questions related to installation and/or
usage in IRC actively.

It is true that Senlin has been initiated by a single company but the project
has now attracted contributors from more than 6 affiliations. We have reasons
to believe the adoption by Big Tent will help encourage more contributions to
its design and implementation, thus improve the diversity status of the team.

[1]https://wiki.openstack.org/wiki/Senlin
[2]https://review.openstack.org/#/q/project:openstack/senlin,n,z
[3]https://review.openstack.org/#/q/project:openstack/python-senlinclient,n,z
[4]https://review.openstack.org/#/q/project:openstack/senlin-dashboard,n,z
[5]http://eavesdrop.openstack.org/irclogs/%23senlin/
[6]https://wiki.openstack.org/wiki/Meetings/SenlinAgenda
[7]http://git.openstack.org/cgit/openstack/senlin/tree/doc/api-site/api-ref/src/wadls/clustering-api/src/v1
[8]http://git.openstack.org/cgit/openstack/senlin/tree/doc/source/getting_started
[9]http://git.openstack.org/cgit/openstack/senlin/tree/doc/source/developer

Change-Id: I3a6581578cbff3e76ed197d2dfecc7a2ef3e1dd5