Freezer application to join the Big Tent
Mission and Vision:
To provide tools for managing backup/restore and disaster recovery.
Full version available here [0].

The temporary PTL is Fausto Marzi.
Starting from April a new PTL will be elected by using the standard
OpenStack mechanism and timeline [1]

The #openstack-freezer IRC channel on freenode is logged and the channel is
notified of such logging [2]. Our public meeting logs are stored [3].

Many design aspect has been defined at the Vancouver Summit [4].

The project uses public code reviews on the OpenStack infrastructure.
A part the first 2-3 months of the project creation,
Freezer has been using the +2/+A process requiring two unique core
reviewers to review a patch.

Our gating runs pep8, pylint and other unittest and integration tests.
In the last weeks a Devstack plugin was also implemented [5].

Currently there is a dedicated Team working on Freezer from one
of the top 3 Company contributors in OpenStack. There are also
external contributors to the project [6].

As this is a new project the number of contributors is not huge,
but there is a strong interest everywhere and the feedback from
the community was always positive and encouraging.

Freezer uses IRC and OpenStack related ML for all the communications.

We are doing all our best to make this project useful and successfull
for the whole OpenStack Community.

Freezer Project repositories:

- https://github.com/openstack/freezer/
- https://github.com/openstack/freezer-api
- https://github.com/openstack/freezer-web-ui

[0] https://etherpad.openstack.org/p/freezer-manifesto
[1] https://wiki.openstack.org/wiki/PTL_Elections_April_2015
[2] http://eavesdrop.openstack.org/irclogs/%23openstack-freezer/
[3] https://etherpad.openstack.org/p/freezer_meetings
[4] https://www.openstack.org/summit/vancouver-2015/summit-videos/presentation/freezer-the-openstack-back-up-as-a-service-platform
[5] https://review.openstack.org/#/c/233683/
[6] http://stackalytics.com/?project_type=openstack-others&release=all&metric=commits&module=freezer

Change-Id: Ice43efe6f1e6691d44730ec38f38479780af49a8