Add assert:supports-upgrade tag to Neutron
Based on requirements defined in [1], I believe that is
a fair statement to claim cold upgrades support for Neutron.

Rolling upgrades are not too far off, but not within our reach
to allow us to claim assert:support-rolling-upgrade, just yet.
This is mainly because of lack of full stack integration testing
with services arranged in a mid-upgrade manner. This is planned
for Mitaka, and I count to reconsider adding the latter tag in
time for the release.

[1] http://governance.openstack.org/reference/tags/assert_supports-upgrade.html#requirements

Change-Id: I691325fe5f8562fb59239ecde75f0992a53d599e