Resolution about Image Uploads and Linux Kernels
It has become clear that it's time for the TC to make a clear statement
about our intent on the point of Images and Kernels.

As an Open Source project, we have gone out of our way to enable
productization and differentiation by our vendors. Our 3rd Party CI
system is massive and Open- and our pluggability is almost cripplingly
comprehensive.

To give up on interoperable image uploads simply because a vendor would
like to do one thing that a little bit different isn't just wrong, it would
be giving up the ENTIRE effort.

We take a stand on very few things. We should take a stand on this.

Change-Id: I41247052240eaec19a9002cbefc52297d0d14566