Refactor 4 opens as in Project team guide
The version of the "4 opens" that was in the wiki had been
refactored when copied to the Project Team Guide, to get
rid of the "we follow this principles" section which was
partially redundant and not really one of the "opens". Some
bits were moved to the corresponding sections, and redundant
text was removed.

Now that the wiki version has been imported in governance,
re-propose this same refactoring for inclusion in the reference
document.

Change-Id: I4da77173ce64b1e64106d48c7d9e3262dde561f8