Deprecate the use of the term "Stackforge"
When this resolution was originally adopted, the TC felt that the
term 'Stackforge' was still useful as a sort of shorthand for
'the area where unofficial projects can develop before they became
official'.  In keeping with that, we have continued to refer to
the hosting space for unofficial projects as "Stackforge".  This
has certainly resulted in people thinking our documentation was
out of date, and has also likely resulted in missed opportunities
to communicate that projects within openstack/ but not approved
by the TC are unofficial, rather than official OpenStack projects.

Let's drop the effort to keep using "Stackforge" in official
contexts.

Change-Id: I70a6dedff3e7905d029b4ea477079ce62d310984