move extra-atcs into projects.yaml
Having multiple file formats for data related to projects is resulting
in duplicated tooling. Having the extra ATC members listed outside of
projects.yaml is also allowing us to have different team names in the 2
files, which leads to other errors. Move the extra-atcs into
projects.yaml and update the rendering code to use the data from the new
location.

Change-Id: I9857af4caf23c18ff7e39cdfbf27330f53f31135