Clarification of licensing requirements
Licensing requirements under the big tent have been unclear at best.
The Foundation bylaws impose some of them, but their interpretation
was left as an exercise to the reader. The TC traditionally came up
with rules for dependencies (through the requirements repository
README) but those were mostly unchecked. Finally, we need flexibility
for the project infrastructure, which should be able to run any
open source solution that fits.

After much back and forth with the Foundation legal counsel on what
could be acceptable under the bylaws, we agreed that the following
proposal was acceptable, and compatible with the current situation.

I propose we add it as a reference document to our governance
repository so that it's easier to point prospective projects to it,
and to track changes in the list of acceptable licenses over time.

Change-Id: I7ed11388e43bd5fcf2e8e6595c353429300387d5