Add python-wsmanclient project under Ironic
The library will be used by the AMT and DRAC drivers in Ironic. They are
using the WSMan protocol for communication. Both used openwsman
previously but the DRAC driver recently moved to its own client
implementation, which is also in the interest of the AMT driver because
of the limitations of openwsman.

Change-Id: If40283a24a6656808e43e5b010222f5b5d80da7e
Depends-On: I431f1607f01c0fedda1ab91b602ed440b2f7508a