Add project Smaug to OpenStack big-tent
Following the directive
http://governance.openstack.org/reference/new-projects-requirements.html

OpenStack Mission Alignment:
Smaug outlines a framework of APIs and orchestration services that
extends OpenStack to enable its users to deploy Backup and Restore
products from multiple vendors, and use them through a unified,
*coherent* interface.

Smaug exposes project-level backup and restore capabilities, which are
driven by various implementations, based on the selection of the
OpenStack instance administrator.

Smaug is designed with multiple cloud use cases in mind, such as private
enterprise deployment that may require tiering of project resource data
protection based on service level (SLA), or public cloud scenarios where
the *provider* admin may wish to provide data protection "as a service",
while the *project* admins expect to determine data protection policies by
themselves.

Smaug was initially launched as a joint-effort between Huawei and IBM
and announced during the OpenStack Mitaka summit in Tokyo. [1]

Our Mission Statement:

1) *Formalize* a unified interface with a unified schema to describe all
   resources, dependencies, policies and actions

2) *Any resource* should be protected, therefore Smaug should be
   extendable and open to any vendor

3) *Diversity* of vendors and solutions is of paramount importance, and
   Smaug should expose all features to the users.

Following the 4 opens
Open Source:
Smaug is 100% open source, everything from implementation
to design to future features and roadmap are happening and shared
openly and presented in various meet-ups and documents.
Smaug plans to become production grade open source OpenStack
Data Protection API framework, to be used with multiple implementations,
both in OpenStack (like Freezer) and outside of OpenStack.

Open Community:
We are working closely with the community to share use cases and problems
and decide on future priorities for the project.
It is our mission to collaborate with as many members/companies
in the community as possible and we welcome any feedback and any desire
to help us shape Smaug's future.
This is done using the mailing list, IRC, and weekly meetings [2].

Open Development:
- All Smaug code is being code reviewed in OpenStack Gerrit [3][4][5]
- Smaug has a core team which openly discuss all issues [6]
- Smaug support gate tests which run unit tests and fullstack tests
- Smaug collaborates other projects such as Cinder to try and find
  optimal integration solutions
- Bugs are managed by OpenStack launchpad [7]

Open Design:
- All designs and specs are published for review and managed in
  OpenStack launchpad
- Smaug conducts a weekly IRC meeting to discuss all designs and future
  roadmap
- Smaug repository offers documentation and diagrams to explain the
  current design, code and future roadmap
- Everything is discussed openly either on the review board,
  Smaug IRC channel (which is logged), or the ML.
- Smaug's mission statement is to become an integral part of OpenStack

See Also:
- Smaug Wiki > https://wiki.openstack.org/wiki/smaug
- vBrownBag Presentation * https://youtu.be/_tVYuW_YMB8
- Documentation Root
    > https://github.com/openstack/smaug/blob/master/doc/source/
- Smaug Source > https://github.com/openstack/smaug
- Smaug Launchpad > https://launchpad.net/smaug
- Smaug Meetings
    > https://wiki.openstack.org/wiki/Meetings/smaug
    > http://eavesdrop.openstack.org/meetings/smaug/

[1] https://youtu.be/6RU_4vZZiLQ?t=15m23s
[2] http://eavesdrop.openstack.org/meetings/smaug/
[3] https://review.openstack.org/#/q/project:openstack/smaug
[4] https://review.openstack.org/#/q/project:openstack/smaug-dashboard
[5] https://review.openstack.org/#/q/project:openstack/python-smaugclient
[6] https://trello.com/b/Sudr4fKT/smaug
[7] https://bugs.launchpad.net/smaug

Change-Id: I323c90f4f6d59ea0603245370f37ccbb46893adb