Add puppet-ovn to Puppet OpenStack
networking-ovn is a project part of OpenStack Neutron:
http://docs.openstack.org/developer/networking-ovn/readme.html

"OVN provides virtual networking for Open vSwitch and is a component of
the Open vSwitch project. This project provides integration between
OpenStack Neutron and OVN."

Red Hat started to work on a Puppet module that installs & configures OVN
services, on top of Open-vSwitch:
https://github.com/redhat-cip/puppet-ovn
We would like to move it to OpenStack under Puppet OpenStack tent.

Why not deploying OVN with puppet-vswitch?
OVN is a part of the OVS project. It's even in the same git repo as OVS
right now.
OVN consists of some new daemons and databases. It's optional components
that you may enable on top of OVS.
But openstack/puppet-vswitch abstracts away the vswitch details in its
interface, so puppet-ovn would likely consume puppet-vswitch.

Change-Id: Idd8dc346b8368c926a5f2fc5c2cc8c34376f30f5