Remove rolling-upgrade tag from ceilometer
This tag was prematurely applied to this project, as it currently
doesn't follow point #4 in the tag process -
https://governance.openstack.org/reference/tags/assert_supports-rolling-upgrade.html

"Full stack integration testing with services arranged in a
mid-upgrade manner is performed on every proposed commit to validate
that mixed-version services work together properly. This testing must
be performed on configurations that the project considers to be its
reference implementations. The arrangement(s) tested will depend on
the project (i.e. should be representative of a
meaningful-to-operators rolling upgrade scenario) and available
testing resources. At least one representative arrangement must be
tested full-stack in the gate."

Ceilometer has a simple cold upgrade grenade job, but it doesn't
demonstrate mixed mode in the gate.

Change-Id: Ie546d032114c519653c7dd2ef738023f633d5026