Add threat analysis for vulnerability:managed
The security team generally believes an audit or review is unfair
because it requires a massive amount of work.  They prefer to do
threat analysis.  The log of the discussion was held in the
security IRC meeting here:

http://eavesdrop.openstack.org/meetings/security/2016/security.2016-03-17-17.00.log.html#l-183

Add threat analysis as an option for the vulnerability:managed tag
as a result of the expert security team's thinking.

Change-Id: I6b3f078b59108238c6b899d86a637b5e88d19aa5