Update some tags for the Trove project
This is the third of multiple changes to fix tags on the trove project.

In reviewing the tags for the Trove project, I find some tags and
assertions to be missing and I have added them. I have run the
validate-tags tool and it reported no errors relative to Trove. (to
save time, I ran the tool with a modified projects.yaml file that had
only Trove). I have retained the release:has-stable-branches on
trove-dashboard as that is the intent.

Change-Id: I8ed71de27e70e8f1f25c9fe9fb1ec973a79355a2