Appointed PTLs for Newton
This change is about appointing PTLs for projects which did not
have candidates in time for the regular PTL election. During
the meeting on March 22nd, the Technical Committee members present
decided to propose the following:

- EC2API: keep Alexander Levine as PTL for the Newton cycle
- Stable maintenance: appoint Tony Breeds for the Newton cycle
- Winstackers: keep Claudiu Belu as PTL for the Newton cycle

If there is disagreement on a specific name, please indicate it in
your comments and I'll split the patch up.

Change-Id: I17906bba073effc18d456cc508dfde3896823870