Update some tags for the Trove project
This is the first of multiple changes to fix tags on the trove
project. In reviewing the tags for the Trove project, I find some tags
and assertions to be missing and I have added them.

I have run the validate-tags tool and it reported no errors relative
to Trove. (to save time, I ran the tool with a modified projects.yaml
file that had only Trove). I have retained the
release:has-stable-branches on trove-dashboard as that is the intent.

I have manually rebased this change on top of 295887 to eliminate the
merge conflict later.

Change-Id: I49f568df804c6a0c2ee83e06647c2bcf2b68f9d0