Update Magnum Description and Mission
At our Newton design summit in Austin, our community reached a
consensus to narrow the scope of the Magnum project to focus on
integrating container orchestration engines (COEs) with OpenStack,
and making them easy to create, manage, and delete in an OpenStack
cloud with full multi-tenancy support.

API features intended to uniformly create, manage, and delete
individual containers across any COE will be removed from Magnum's
API, and may be re-introduced as a separate project with a new name.

This change is justified by the recognition that we have two distinct
groups of stakeholders who value different goals for different reasons.
Conflation of refining a common COE abstraction, and deep integration
of each COE for native API deployment/management within Magnum has
proven to be problematic. We will solve this conflation by clarifying
a narrow scope for Magnum, and allowing another project to pursue a
common COE abstraction. This allows each group of stakeholders to focus
on their mutual interests.

Change-Id: I6682e78827e168a517d3d91e536f9d3fd0ec6c6a