Adds a docs: URL entry to projects.yaml
contributor: <UR>L - These are docs for how to contribute to a project listed, meant
for contributors to understand and work on the project. May also contain user
docs.

api: <URL> - These are docs for REST API reference information for service projects.

A screenscrape of the OpenStack projects doc page gives a listing
of contributor docs.[2]

Annotated per-repo.

[1]http://docs.openstack.org/developer/openstack-projects.html
[2]http://paste.openstack.org/show/515989/

Change-Id: I170b1f5cc2bcef4f9c5dc4fd25deb4709339acaa