Revisit the Neutron list of repos
LBaaS, FWaaS, VPNaaS and Dynamic Routing are listed as
part of the Neutron deliverable. These components have
always been meant to be optionally deployed, and like
the other deliverables, they should be described
individually.

As a result, they each get their own tags depending
on the level of compliance with the tag definition.

Change-Id: I3e86210ffbeceec8783b58120367fb8cd8f162dc