Remove Cue project team
The Cue project team activity level has dropped below what is
expected of OpenStack official project teams, so it should be
removed and become an unofficial project. In particular:

* There was no release of Cue during the Mitaka cycle
* Only peripheral changes were pushed in the last 3 months
* The Cue PTLs did not answer to questions about Design Summit
  space in Austin, and therefore did not get any space there
* There were no commits or reviews since Apr 16, 2016
* Last 3 messages about Cue on the ML were:
  - The message asking for Cue team inclusion in June, 2015
  - The PTL candidacy in September 2015
  - The PTL candidacy in March 2016
* Last Cue team meeting on IRC was in December 2015

The Cue project repositories will continue to live on OpenStack
Infrastructure, so if it gets back some activity we would
definitely consider re-adding them. But at this point keeping them
in the big tent is doing more harm than good.

Change-Id: I4822415a4ee18a4c52f9ba149b0cc36f363d80dd