Add 'level playing field' requirement
From an upstream perspective, I see us as being in the business
of providing open collaboration playing fields in order to build
projects to reach the OpenStack Mission. We collectively provide
resources (infra, horizontal teams, events...) in order to enable
that open collaboration.

An important characteristic of these open collaboration grounds
is that they need to be a level playing field, where no specific
organization is being given an unfair advantage. I expect the teams
that we bless as "official" project teams to operate in that fair
manner. Otherwise we end up blessing what is essentially a trojan
horse for a given organization, open-washing their project in the
process. Such a project can totally exist as an unofficial project
(and even be developed on OpenStack infrastructure) but I don't
think it should be given free space in our Design Summits or
benefit from "OpenStack community" branding.

So if, in a given project team, developers from one specific
organization benefit from access to specific knowledge or hardware
(think exclusive access to proprietary hardware or software that
the open source code primarily interfaces with), then this project
team should be rejected under the "open community" rule.

This requirement (like all requirements in this document) is meant
to be interpreted by humans on a case-by-case basis. In particular,
there is a grey area around drivers for proprietary solutions that
requires human judgment here. As a rule of thumb, if the open
implementation was unusable open core bait to lure people into
using the one and only proprietary driver, it would be a problem.
If the open implementation was fully functional and nothing
prevented adding additional proprietary drivers to the code base,
there would be no problem.

Change-Id: I0cd689afb8ddb7a269b9ce378e4c6647652f6977