remove release:managed tag
The release:managed tag used to describe projects fully managed by the
release management team. With the new automation and changes to our
release processes, we have expanded the list of projects that are
"managed" in this sense to all official project teams using cycle-based
release models. This tag has therefore become redundant, and can be
removed.

Change-Id: I51a242c2c706fd3040af0dea391175ef5206aa6b
Signed-off-by: Doug Hellmann <doug@doughellmann.com>